i2c_spi_buses.h: work around astyle version inconsistency
When running make format with astyle 3.1
(which is installed with the setup scripts)
the style gets corrected the expected way but
CI seems to use an older version that parses the
bracketing wrong and corrects it again.

To work around that I put a newline in between.
This mutes the constantly failing style checks.